chore: remove gpuCards mocks
