+++ packages/cube-frontend-web-app/src/mocks/gpu.ts
@@ -1,3 +1,9 @@
+import { delay, http, HttpResponse } from 'msw'
+import {
+  ListNodeGPUCardsResponse,
+  UpdateNodeGPUCardPutRequest,
+  UpdateNodeGPUCardResponse,
+} from '@cube-frontend/api'
 import {
   GPUCardStatus,
   GPUResourceType,
@@ -575,3 +581,36 @@ export const mockGpuResource: ListNodeGPUCardsResponseDataInner[] = [
     attachedInstances: [],
   },
 ]
+
+export const mockListNodeGpuCards = http.get(
+  '/api/v1/datacenters/:dataCenter/nodes/:nodeName/gpuCards',
+  async () => {
+    await delay(2000)
+    const res: ListNodeGPUCardsResponse = {
+      code: 200,
+      data: mockGpuResource,
+      msg: 'Success',
+      status: 'success',
+    }
+
+    return HttpResponse.json(res)
+  },
+)
+
+export const mockUpdateNodeGpuCard = http.put<
+  { dataCenter: string; nodeName: string; gpuId: string },
+  UpdateNodeGPUCardPutRequest,
+  UpdateNodeGPUCardResponse
+>(
+  '/api/v1/datacenters/:dataCenter/nodes/:nodeName/gpuCards/:gpuId',
+  async () => {
+    await delay(2000)
+    const res: UpdateNodeGPUCardResponse = {
+      code: 200,
+      msg: 'Success',
+      status: 'ok',
+    }
+
+    return HttpResponse.json(res)
+  },
+)
